🚚structure: QnaUserResult 컴포넌트 파일명 변경 및 이동
