--- src/components/QnADetailResult/index.ts
@@ -1,1 +1,0 @@
-export { default as QnADetailResultCard } from './QnADetailResultCard';
